// libs/enums/notification.enum.ts
export enum NotificationType {
	CONTACT_AGENT = 'CONTACT_AGENT',
	COMMENT_CREATED = 'COMMENT_CREATED',
	LIKE_CREATED = 'LIKE_CREATED',
	FOLLOW_CREATED = 'FOLLOW_CREATED',
	ADMIN_NOTICE = 'ADMIN_NOTICE',
	/* Mirrors apps/gotrip-api/src/libs/enums/notification.enum.ts */
	GUIDE_REQUEST = 'GUIDE_REQUEST',
	GUIDE_APPROVED = 'GUIDE_APPROVED',
	GUIDE_REJECTED = 'GUIDE_REJECTED',
	MESSAGE_RECEIVED = 'MESSAGE_RECEIVED',
}

export enum NotificationStatus {
	WAIT = 'WAIT',
	READ = 'READ',
	DELETED = 'DELETED',
}

export enum NotificationGroup {
	MEMBER = 'MEMBER',
	TOUR = 'TOUR',
	ARTICLE = 'ARTICLE',
	COMMENT = 'COMMENT',
	NOTICE = 'NOTICE',
}